Associating posts with users

+++ blog/posts/views.py
@@ -1,6 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from urllib.parse import quote
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -8,9 +7,13 @@ from django.core.paginator import Paginator
 
 
 def posts_create (request):
+    # to make sure the user is an admin or a staff member
+    if not request.user.is_staff or not request.user.is_superuser:
+        return Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfully created")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -20,9 +23,7 @@ def posts_create (request):
 
 def posts_detail(request, slug):
     object = get_object_or_404(Post,slug=slug)
-    # url encoded text
-    sharable = quote(object.content)
-    context = {'title': object.title, 'object': object, 'sharable': sharable}
+    context = {'title': object.title, 'object': object}
     return render(request, 'post_details.html', context)
 
 
@@ -37,6 +38,8 @@ def posts_list(request):
 
 
 def posts_update (request, slug):
+    if not request.user.is_staff or not request.user.is_superuser:
+        return Http404
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance= instance)
     if form.is_valid():
@@ -49,6 +52,8 @@ def posts_update (request, slug):
 
 
 def posts_delete(request, ID):
+    if not request.user.is_staff or not request.user.is_superuser:
+        return Http404
     instance = get_object_or_404(Post, id=ID)
     instance.delete()
     messages.success(request, "DELETED")
